requirements.txt
Pillow
django-bootstrap4
django-crispy-forms
pillow django-imagekit
folium
googlemaps